Update Agriculture.md
Very minor edits:
line 11 - added 's' "Many package(s) in base R..."
line 33 - adjusted tab at start to be consistent with others
line 188 - 'an' should be 'can' in "that an be expanded to multiple..."
line 206 - changed capitalization phenorice model to PhenoRice
line 212 - removed capitalization of the 'd' in de Wit
line 249 - added space after "pathotypes." and before "Evolution of resistance.."
line 273 - removed ',' after Choi to make consistent with other citations
line 275 - removed 's' from 'packages' in (note this package is optimized)
line 275 - usually you have highlighted in blue a clickable link with the author(s) and year, Litchefield and Wilcoxon is not all highlighted. It's very minor and I wasn't sure how to change this. There are a few others but links on the years works so I wasn't too concerned.